split demo app into library and binary
- Ensure that `reflectapi.json` is also updated when running tests.
