+++ reflectapi-demo/src/tests/mod.rs
@@ -8,3 +8,14 @@ mod serde;
 mod test_lib;
 
 pub(crate) use assert::*;
+
+#[test]
+fn write_schema() {
+    let (schema, _) = crate::builder().build().unwrap();
+
+    std::fs::write(
+        format!("{}/{}", env!("CARGO_MANIFEST_DIR"), "reflectapi.json"),
+        serde_json::to_string_pretty(&schema).unwrap(),
+    )
+    .unwrap();
+}
