+++ reflectapi-demo/tests/run/main.rs
@@ -1,0 +1,2 @@
+#[test]
+fn test() {}
